app1/requirements.txt
Authlib==1.2.0
blinker==1.6.2
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cryptography==40.0.2
Flask==2.3.2
idna==3.4
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pkg_resources==0.0.0
pycparser==2.21
pyOpenSSL==23.1.1
requests==2.31.0
urllib3==2.0.2
Werkzeug==2.3.4
zipp==3.15.0
